Skip orphaned SMS messages. Fixes issue 8
I've determined that phantom messages can show up when you delete a
conversation in the messaging app but for some reason the app
accidentally doesn't delete all messages in that conversation. So you're
basically left with stray messages that aren't part of a conversation.
The messaging app only shows conversations at the top-level, not
messages. DashClock was seeing those stray messages and showing them.
Now, it first collects a list of all active conversations, and then
checks to make sure unread messages are part of an active conversation
before showing it.